added "task" api routes

+++ models/task.go
@@ -7,7 +7,7 @@ import (
 type Task struct {
 	ID          uint   `gorm:"primaryKey"`
 	Description string `gorm:"type:varchar(255)"`
-	TaskOrder   uint   `gorm:"unique"`
+	TaskOrder   int64  `gorm:"unique"`
 	Checked     bool
 	CreatedAt   time.Time
